refactor(monitoring): change check_interval type from int to float in monitoring
The `check_interval` parameter in `watch_pid_file` and `monitor_mover_lifecycle` was previously an integer, implying whole second intervals. This commit refactors it to a float to allow for sub-second precision, enhancing the responsiveness and configurability of the monitoring functions.

In the test suite, various `asyncio.sleep` durations and `timeout` values have been significantly reduced to align with the new sub-second precision, making the tests run faster and more efficiently. For instance, `timeout` values in `validate_process_with_timeout` and `asyncio.timeout` blocks were adjusted from `1.0` or `2` to `0.1` or `0.05`. Similarly, `asyncio.sleep` calls in test helpers were changed from `0.5` or `1.5` to `0.01` or `0.03` to reflect finer-grained timing. The `check_interval` parameter in `test_watch_respects_different_intervals` was also changed from `[1, 2, 3, 5]` to `[0.01, 0.05]`.

+++ src/mover_status/core/monitoring.py
@@ -551,7 +551,7 @@ class MoverLifecycleStateMachine:
 async def watch_pid_file(
     pid_file_path: Path,
     *,
-    check_interval: int = 1,
+    check_interval: float = 1.0,
 ) -> AsyncGenerator[PIDFileEvent, None]:
     """Watch PID file and yield events on state changes.
 
@@ -729,7 +729,7 @@ async def watch_pid_file(
 async def monitor_mover_lifecycle(
     pid_file_path: Path,
     *,
-    check_interval: int = 1,
+    check_interval: float = 1.0,
 ) -> AsyncGenerator[MoverLifecycleEvent, None]:
     """Monitor mover process lifecycle and yield state transition events.
